requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
openai
python-dotenv
streamlit
transformers
torch
rake_nltk